requirements.txt
numpy
scipy
Pillow
pygame